refactor(design-tokens): share the control helpers the other blocks need
`useLinkedMeasureState` and `resolveColorLiteral` both lived under
`src/blocks/singlebtn/`, but neither is block-specific: the first derives the
linked/individual mode for any responsive four-slot measure control (its own
docblock says "and any control sharing that shape"), and the second reads a
token's CSS variable off `document.documentElement`, which
`palette-swatch-preview.js` stamps with the selected block's effective palette
regardless of which block that is.

Move them beside the machinery they belong to -- the hook into
`token-indicators`, the color resolver into `design-tokens` next to
`useColorGroups`, which supplies the other half of a `ColorControl`'s data --
so a block wiring either reaches for one public entry point rather than
importing across into the Button.

Export `BorderControl` and `BoxShadowControl` from `token-controls`' index too.
The library README calls anything it does not export internal, yet both were
already consumed by two hosts through a deep path; the editor adapters now use
the public import.

Docblock corrections that follow from the move: the three `Editor*Control`
wrappers no longer claim they only ever render inside the Button's sidebar, and
`EditorShadowControl` describes the native shadow shape as the one every
shadow-carrying block shares rather than as the Button's own.

No behavior change.
